storage: move catalog scan into CatalogSnapshot::build(&mut NodePageStore)
CatalogSnapshot::build now takes &mut NodePageStore directly and uses
scan_leaf_cells + CellReader to traverse the catalog tree. This removes
the dependency on BTree/Cursor for cache construction.

In btree.rs: build_catalog_snapshot is deleted; ensure_catalog_cache
becomes a one-liner calling CatalogSnapshot::build. CATALOG_ROOT is
defined in catalog_cache.rs as PageId(1); btree.rs keeps its own u32
constant for use with BTree::open.
